+++ p_pspr.h
@@ -60,7 +60,7 @@ typedef enum
 typedef struct
 {
     state_t*	state;	// a NULL state means not active
-    int32_t		tics;
+    int16_t		tics;
     fixed_t	sx;
     fixed_t	sy;
 
@@ -70,11 +70,11 @@ typedef struct
 typedef struct
 {
     ammotype_t	ammo;
-    int32_t		upstate;
-    int32_t		downstate;
-    int32_t		readystate;
-    int32_t		atkstate;
-    int32_t		flashstate;
+    int16_t		upstate;
+    int16_t		downstate;
+    int16_t		readystate;
+    int16_t		atkstate;
+    int16_t		flashstate;
 
 } weaponinfo_t;
